Conflicts:
	xray-svn-trunk/xrCore/Xr_ini.cpp
	xray-svn-trunk/xrCore/xrCore.h
	xray-svn-trunk/xr_3da/FBasicVisual.h
	xray-svn-trunk/xr_3da/device.h
	xray-svn-trunk/xr_3da/xr_collide_form.cpp
	xray-svn-trunk/xr_3da/xr_collide_form.h

+++ xray-svn-trunk/xr_3da/xrGame/Bolt.cpp
@@ -54,7 +54,7 @@ bool CBolt::Useful() const
 	return false;
 }
 
-bool CBolt::Action(s32 cmd, u32 flags) 
+bool CBolt::Action(u16 cmd, u32 flags) 
 {
 	if(inherited::Action(cmd, flags)) return true;
 /*
