+++ xray-svn-trunk/xr_3da/xrGame/HudItem.h
@@ -12,6 +12,7 @@ struct HUD_SOUND;
 class CInventoryItem;
 
 #include "actor_defs.h"
+#include "inventory_space.h"
 #include "weaponHUD.h"
 
 class CHudItem {
@@ -40,7 +41,7 @@ public:
 	virtual void	StopHUDSounds		()				{};
 	
 	//для предачи команд владельцем
-	virtual bool	Action				(s32 cmd, u32 flags);
+	virtual bool	Action				(u16 cmd, u32 flags);
 	virtual void	onMovementChanged	(ACTOR_DEFS::EMoveCommand cmd)				{};
 
 	virtual void	OnDrawUI			()				{};
