+++ xray-svn-trunk/xr_3da/xrGame/Inventory.h
@@ -75,7 +75,7 @@ public:
 	PIItem					ActiveItem			()const					{return m_iActiveSlot==NO_ACTIVE_SLOT ? NULL :m_slots[m_iActiveSlot].m_pIItem;}
 	PIItem					ItemFromSlot		(u32 slot) const;
 	void					ActivateNextItemInActiveSlot();
-	bool					Action				(s32 cmd, u32 flags);
+	bool					Action				(u16 cmd, u32 flags);
 	void					Update				();
 	// »щет на по€се аналогичный IItem
 	PIItem					Same				(const PIItem pIItem, bool bSearchRuck) const;
@@ -184,5 +184,5 @@ protected:
 
 	bool				m_drop_last_frame;
 
-	void				SendActionEvent		(s32 cmd, u32 flags);
+	void				SendActionEvent		(u16 cmd, u32 flags);
 };
